requirements.txt
google-cloud-storage
numpy
scipy
networkx
matplotlib
tqdm
